re-organize redtent controller/wizard
